requirements.txt
fastapi==0.85.1
uvicorn==0.16.0
PyJWT==2.6.0
python-decouple==3.6
passlib[bcrypt]==1.7.4